refactor: pagesを切り出し・Pageとコンポーネントとは区別
