+++ src/Article.ts
@@ -1,9 +1,8 @@
 import { Tags } from './Tags';
 import dayjs from './dayjs';
 import { Post } from './post';
-import { Page } from './render';
 
-const Article = (post: Post): string => `
+export const Article = (post: Post): string => `
   <article>
     <header class="bl_text_header">
       <time class="bl_text_date">
@@ -23,13 +22,3 @@ const Article = (post: Post): string => `
       ${Tags(post.tags)}
     </footer>
   </article>`;
-
-export const ArticlePage = (post: Post): Page => ({
-  body: Article(post),
-  suffix: ` :: ${post.id}`,
-  description: `${post.plainText.substr(0, 110)}…`,
-  title: post.title
-    ? `${post.title}｜placet experiri :: ${post.id}`
-    : `placet experiri :: ${post.id}`,
-  archiveHeader: '',
-});
